requirements.txt
FastAPI
pandas
numpy
uvicorn
httpx
python-dotenv
pydantic
sqlalchemy
sqlite-utils
sentence-transformers
faiss-cpu
pdfplumber
pymupdf
uuid
tqdm
pytest
langchain
openai>=1.0.0